Telemetry snitching despite settings
This code does not respect the users telemetry settings.

It specifically calls a telem service with detailed user information, os information and build info, without consent.

This was specifically added by the void contributors and is not vestigial code. It was added very recently, after void telem settings were implimented.

This was intentional. This user should be investigated for conflicts of interest.